pushing some other changes

+++ profiler/java/yuca/profiler/YucaProfiler.java
@@ -45,7 +45,7 @@ public final class YucaProfiler implements Profiler {
     }
   }
 
-  private static int getNormalFrequency() {
+  private static long getNormalFrequency() {
     long[] freqs = CpuFreq.getSetFrequencies();
     return freqs[(int) Math.floor(freqs.length / 2)];
   }
